fixed failing check
updated len variable to return 0 when string is NULL

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -12,8 +12,8 @@
 
 size_t print_list(const list_t *h)
 {
-	size_t len;
-	size_t elem_count;
+	int len;
+	int elem_count;
 	char *string;
 
 	elem_count = 0;
